+++ src/logger.js
@@ -1,0 +1,16 @@
+/**
+ * Nextcloud - Ownpad
+ *
+ * This file is licensed under the Affero General Public License
+ * version 3 or later. See the COPYING file.
+ *
+ * @author Olivier Tétard <olivier.tetard@miskin.fr>
+ * @copyright Olivier Tétard <olivier.tetard@miskin.fr>, 2024
+ */
+
+import { getLoggerBuilder } from '@nextcloud/logger'
+
+export const logger = getLoggerBuilder()
+	.setApp('ownpad')
+	.detectUser()
+	.build()